Refactor: Extract DarkGlassStatCard components
- Extracted `DarkGlassStatCard`, `DarkGlassStatCardPainter`, and `BaseDarkGlassStatCard` from `FoodDetailScreen.kt` into their own files under `presentation/components/card/dark_glass_card/`.
- Updated `FoodDetailScreen.kt` to import and use the new `DarkGlassStatCard` and `DarkGlassStatCardPainter` composables.
- Updated `MeadDetailScreen.kt` to use the new `DarkGlassStatCard` and `DarkGlassStatCardPainter` composables.
- Replaced `FoodImage` composable in `FoodDetailScreen.kt` and `MeadDetailScreen.kt` with `SmallFramedImage`.
- Corrected a typo in a log message within `MeadDetailViewModel.kt` (FoodDetailViewModel -> MeadDetailViewModel).
